Update patch set 2
Patch Set 2: Code-Review+2 Verified+1

Re-approving.

I think the use cause you're thinking is not a problem. In my opinion the cost model is the one responsible for returning the equivalence classes to which we want to add arcs. It's fine if it's only a subset of task's typical equiv classes as long as the cost model maintains the info that the task has other equiv classes.

Patch-set: 2
Reviewer: Ionel Corneliu Gog <1005213@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2
Label: Verified=+1