Fixed simulator's AddTask handling of duplicate submission and tasks.
This change updates AddTask to correctly remove the state added by
a duplicate task that gets ignored. With a good hash function there
should only be a duplicate job in the Google trace. Hence, it is safe
to simply ignore the job and make sure to remove the state we've
created for the job.

Moreover, this change makes sure that we correctly handle the task
failures from the trace. In the trace, a task can be submitted again
if a task is killed, lost, evicted or fails. Unfortunately, we can't
replay these events exactly because they're affected by the scheduling
decisions we make. Instead the compute the runtime of the task as the
total time it was running, including runs in which it failed. In this way,
the task ends up keeping resources busy for the same amount of time
in the simulation.

In order to avoid scheduling a task multiple times we were previously
simply checking trace_task_id_to_td map to see if we've already
scheduled the task. However, this is not enough because we can
have a case in which a task finishes successfully and only afterwards
we get an additional submit event. The change fixes this bug by
storing the trace ids of the taks that have been already submitted
for the entire duration of the simulation.

Affected modules: sim

Change-Id: Ib82b6ed185a946b3d11df90db5ebc3075d5fe6c4