Updated TraceGenerator to print machine id for all relevet task events.
Previously, we were only printing the machine id column of the
task events table when we were scheduling a task. Now, we're
printing it for all the events that have a machine id (i.e.,
EVICT, FAILURE, KILL, SCHEDULE, EVICT). This was possible by
moving the class to trace_generator from FlowGraphManager to
EventDrienScheduler. EventDrivenScheduler is the place where
the calls should be made because it's a scheduler agnostic
class.

Affected modules: misc scheduling scheduling/flow scheduling/simple

Change-Id: Ie7eaee2f98ce552f7be42977b8c6b16f3fe704e9