Added support for task preemption.
This change adds scheduler-side support for task preemption.
Whenever we place a task we add a running arc from the task to
the resource on which the task executes. In a sensible cost model
the running arc will have a cost smaller than all the other
outgoing arcs.

Moreover, the change makes sure that the preemption and
continuation costs are regularly updated.

Affected modules: scheduling/flow

Change-Id: I60b0fb167d6e6271d3ae59646fd02c465333e1e2