Added flag to allow usage of different solver for first solver run.
This change adds support for using a different solver on the
first solver run. This is useful for simulation runs when running
Relax from scratch takes a lot of time. Instead, we can first run
cost scaling and then switch to relax after the first run.

Affected modules: scheduling/flow

Change-Id: Iaf35184129a54c3092c6d06a71ce6104d31f72e2