Changed fork with vfork when forking solver process.
Simulations using CS2 and the CoCo cost model were failing
because just before calling fork(), the memory consumption of
the simulation was higher than half of the machine's memory
capacity. While fork() implements copy-on-write, it probably
still has a check to make sure that there's enough free memory
to copy to if need be. I think this is why the call to fork
was failing with a "not enough memory available" message.

Replacing fork() with vfork() solver the issue. Vfork() doesn't
copy the page tables and likely doesn't have fork()'s safety check.

Affected modules: misc

Change-Id: I83e7f11d2991199b70fea5a8a82b73588648317d