Fixed EquivClassToResourceNode to return the correct capacity type.
EquivClassToResourceNode and TaskFitCount were returning the capacity
as an int64_t despite the capacity's type being uint64_t. Moreover,
TaskFitCount was initializing the number of tasks that fit to
MAX_INT64 which would cause the solver to overflow because several
arcs with max capacity can point to a node.

This change makes sure that FlowGraphManager doesn't overwrite
the return we get from EquivClassToResourceNode. Previously,
FlowGraphManager would set the capacity to the number of tasks slots
when EquivClassToResourceNode was returning a negative capacity. This
is dangerous to do because it can change the behaviour of some cost models.
The cost models should return the correct capacity.

Lastly, the change moves the logic that computes the number of tasks that
fit into a resource subtree to the cost models.

Affected modules: scheduling/flow scheduling/flow/sim

Change-Id: If71d453fe8e1d6a8b779400d3eebbb0eb740954f