Updated WhareMap to not return MAX_INT as a cost.
Returning a MAX_INT cost could cause one of the solvers to overflow.
The change updates the code to bound the arc costs to the value of
the flow_max_arc_cost flag. This will make sure the costs don't overflow,
unless large values are used for the flag.

Affected modules: scheduling scheduling/flow

Change-Id: If354cc53a95e5a99c70385c0bdcadd576fa9a3d9