Updated GetEquivClassToEquivClassesArcs to only return destination ECs.
Previously, for a given EC the method was returning both the source ECs
of the incoming arcs and the destination ECs of the outgoing arcs.
However, the source ECs of the incoming arcs are not required
in the new single pass update implementation. Thus, the change updates
the method to return only the destionation ECs. As a result,
all the methods in the CostModelInterface return only ECs/ResIDs that
are at destinations of the outgoing arcs.

Moreover, the change updates FlowGraphManager to add to the graph the
new EC nodes returned by GetEquivClassToEquivClassesArcs.

Affected modules: scheduling/flow scheduling/flow/sim

Change-Id: Ie749b8c75ef95bd91a9a4c19e58549dbaf15c3f7