Fixed RemoveInvalid*Arcs* to only remove arcs of a compatible type.
There's 2 versions of this method: one that receives a vector of
resources and another one that receives a vector of ECs. Both
methods were removing all the arcs that were not pointing to
an entity present in the vector received as an argument.

This change updates the code so that the methods only remove arcs that
point to nodes not present in the given vector and that are of the same
type as the vector's elements (i.e. arcs to ECs that are not in the
given vector with ECs).

Affected modules: scheduling/flow

Change-Id: I7de7bf0ac67f19a7458f32fdbbd35401103bb545