Properly fix /proc/meminfo reading.
This previous fix (66883b74276a45c496438fdf4837f76f8e1e42b7) did
not actually fully address the issue introduced by the CHECKs in
the code that reads /proc/meminfo. This change replaces the CHECKs
with conditionals that skip the line if it does not match the
expected format.

Affected modules: platforms/unix

Change-Id: I36544fa0b67a17186e6909ffbe49e9922015489d