Moved NodeBindingToSchedulingDelta from SolverDispatcher to FlowGraphManager.
NodeBindingToSchedulingDelta was previously located in SolverDispatcher
despite not accessing any of SolverDispatcher's field. Moreover, the
method requires access to the FlowGraph. This change moves the method
to the FlowGraphManager. In order to be able to do that, I had to
change FlowGraphNodes to store TaskDescriptors rather than TaskID_ts.

Affected modules: scheduling scheduling/flow

Change-Id: I647a60837da0ed27c0bfdac478cef1c661f9dc91