Updated FlowGraphManager to use FlowGraphChangeManager.
Previously, the graph changes were managed by the FlowGraphManager.
As a result, we had plenty of duplicate unsafe code in which we
would directly change FlowGraph* or generate changes. This diff
moves that logic to FlowGraphChangeManager.

With the new design we will also be able to easily implement
several algorithms to reduce the number of DIMACS changes we
generate.

Affected modules: scheduling/flow

Change-Id: Ib027c687caa2b33481bd566953b65a1a13378055