Refactored FlowGraphManager.
This change rewrites the FlowGraphManager to only
traverse the graph nodes from task nodes to sink nodes.
Moreover, the change removes almost all AddOrUpdate methods
and switches to a structure in which we have a method
for every type of update.

As a result of this change we can now also support
dynamic arcs or ECs preferences.

Finally, the change adds extensive unit tests for the
FlowGraphManager.

Affected modules: scheduling/flow

Change-Id: I25698c069f834f77d71688be4dc84f4c64a75f2c