Make TaskLib more robust to unexpected environment
This change adds some sanity checks in TaskLib that make it fail more
usefully when the expected environment variables and/or flags are not
present. Normally, the executor sets these things up, but if people try
to run the example binaries outside Firmament (or the executor code
changes), this will lead to more easily debuggable errors. (Previously,
we would get an exception from somewhere deep in the standard library
because TaskLib was sloppily initialising strings from getenv(3), which
can return NULL.)

Issue: #34
Affected modules: engine examples

Change-Id: I4986494baae6c00f0ff95ffb70c16768f252d179