Out-of-tree build in Travis-CI.
This change moves the Travis builds to out-of-tree builds, since
otherwise there is a name clash for "examples/hello_world" and
similar names that exist both as directories and as targets.

This is a stopgap commit to fix the build; we will want to avoid such
name clashes in the future, so that in-source builds remain possible.

Change-Id: I81d4090d6f7d2d66d18d61e09f547342adcaf593