Build Python protobufs by default with coordinator.
In d881e062d5ba75be481c1023359369e518d940c4, we introduced a CMake
build target for the job submission script's Python protobufs.
However, this target was not built by default because nothing
depended on it. This change adds a dependency between the
coordinator and the scripts target, so that the job submission
script is always built when the coordinator is built.

Issue: #40
Affected modules: build

Change-Id: I01697164057791db7df0b1a29339b6fc1455de6b