Fixed code that computes capacity between resource nodes.
Previously, we were setting the capacity of incoming arcs
to a resource node to the number of free slots. However,
this is not going to work well in the cases when we run
with preemption enabled. This change, updates the code to
set the capacity to num_slots_below when preemption is
enabled.

Affected modules: scheduling/flow

Change-Id: I4ed5cc21eb93f080088d8aa5d0f6564c045c8bd4