Updated EquivClassToEquivClass method to also return the capacity.
Previously, the method was only returning the cost. As a result,
we would have to compute the capacity of the arc connecting the
two ECs outside of the cost model. This change updates the code
so that the method returns the capacity as well.

NOTE: The change breaks the WhareMap cost model.

Affected modules: scheduling/flow

Change-Id: I04d6458fafc7ff4c3ed6aab99034c896ff9edc1e