Implemented Quincy cost model.
The previous QuincyCostModel was broken. This change fixes it and
extends it to faithfully implement the Quincy scheduler described
in the paper.

Moreover, the change makes usage of the DataLayerManagerInterface
so that we can run both real-world workloads and simulation
experiments using the same cost model.

Affected modules: misc scheduling scheduling/flow sim sim/dfs

Change-Id: I208fd89a7e7dc825706ead218ba191ebffc4bac7