Use const references in HTTP UI.
This change moves the HTTP UI to pass pointers into libpion via const
references. This may seem somewhat counter-intuitive, but is actually
what the templates for Boost Bind expect (and it's correct: the pointer
is const, but the object pointed to is not).

The reason this came up is that it has led to build problems with Boost
1.58 on Ubuntu 15.10, even when using gcc as required by #36.

The change also removes the NOLINT comments from the affected lines,
because we now no longer violate the linting script's rules.

Affected modules: engine

Change-Id: I6f1fe720603d1d90cca9e172ecc12aaac2adcd78