Added logic to Quincy to update prefered lists upon machine addition or removal.
Previously, we would not change the preference list and the costs
when a machine was added or removed. This change adds that
feature.

Moreover, the change makes sure we don't count a block twice
even if it is replicated several times on the same machine or
rack.

Affected modules: scheduling scheduling/flow

Change-Id: Iaf9831a41455ab826ada8a9aa770867f4e4de277