Controlled inter-arrival time in synthetic traces.
This change adapts the synthetic trace loader component of the simulator
to take a configurable job inter-arrival time rather than a rate of jobs
per second (which could not be less than one previously).

In addition, this change fixes an issue in the synthetic trace
generation that led to the first job only arriving after the first
inter-arrival time interval had elapsed.

Affected modules: sim

Change-Id: I628a508d749466902bb974f5a7aea01ff9b2c26b