Fix simulation time bug that occurs in batch mode.
This change fixes a somewhat insidious bug that affects simulations
executed in batch mode (i.e., with the flow scheduler running every N
time units, as opposed to as soon as it can). Previously, we would
correctly avoid updating the simulation timestamp in the simulator code,
but would still replay events that occurred during the scheduler runtime
because the OnSchedulingDecisionsCompletion() callback was invoked with
an argument of the the base timestamp (in simulated time) plus the real
scheduler runtime. As a result, we incorrectly replayed events before
applying scheduling decisions if the batch mode frequency was lower than
the scheduler runtime (e.g., --batch_step=1000 and a 100000us scheduler
runtime).

The fix is to pass both the base timestamp and the scheduler runtime
separately, and to only call ProcessSimulationEvents() if not in batch
mode. Passingn the two arguments makes it possible for us to later add
another code path that uses the base timestamp in batch mode. We have to
add the condition inside OnSchedulingDecisionsCompletion() because the
--batch_step flag value is not available outside the simulator code
(and, specifically, not in FlowScheduler).

Affected modules: scheduling scheduling/flow sim

Change-Id: Ic96c0ee7a8b831770e0a49ee70aecc410de18d0d