Add flag for web UI template path.
This change adds a flag (--http_ui_template_dir) that allows the path to
the directory holding the template files for the coordinator's web UI to
be specified. Previously, we were hard-coding paths relative to the
Firmament root directory. Instead, the flag can now be used to specify
relative or absolute paths.

Affected modules: engine

Change-Id: I0ebb3960849510bf5b49187d90de576daa67f4c8