Fixes in the QuincyCostModel.
This change makes several fixes to the QuincyCostModel:
1) It makes sure that we do not try to update the costs to
a rack after we remove it.
2) It makes sure that a task is only removed from the cost model
after we have made to the cost model all the calls related
to the task we're removing.
3) It adds functionality to compute the transfer cost to a machine
that is not in the preffered set. We need to use this functionality
when we schedule a task via the cluster aggregator. Once we've
scheduled the task we compute the continuation cost with requires
the transfer cost to the resource on which the task has been placed.

Affected modules: scheduling/flow

Change-Id: I1c0f5e624fbce326dd53ace9a16c120a2ff72b2a