Fixed computation of machine preference and costs to the cluster AGG.
This change fixes the code that computes the thresholds over which
preference arcs to machines get added. Moreover, the change makes sure
that the cost from a task to the cluster AGG is equal with the cost
to the worst machine and not just to the worst machine from a
preferred rack.

Affected modules: scheduling/flow

Change-Id: I6fe72b901ed53e98b350e35873a0d81f9a1dc87b