Fixed rack EC id creation in QuincyCostModel.
Previously, we were using the size of rack_to_machines map
in order to generate a "unique" rack EC id. However, in
another part of the code we were also deleting entries from
rack_to_machines which means that we could end up creating
racks with the same id.

This change fixes this limitation by introducing a field
that is incremented whenever we create a new rack EC.

Affected modules: scheduling/flow

Change-Id: Iebb44479b511efb258139eab0063c6ee03f718f6