Moved out of Quincy the logic that assigns machines to racks.
The logic should not be in the QuincyCostModel because the
assignment of machines to racks is not Quincy specific. For now,
I've moved the logic to the DataLayers. However, we may want to
implement a ClusterTopology class in the future.

Moreover, the change makes sure that the DataLocations returned
by the DataLayers also contain the id of the rack in which the
block is.

Affected modules: scheduling scheduling/flow sim/dfs storage

Change-Id: Ib331b2425e552fc1d77e9e30ce3f926100c64d0d