Simulator correctly sets task start_time on first solver run.
Previously, the start_time was set to the end time of the first solver
run. However, this could have led to newer tasks having smaller start
times if the second solver runtime is smaller than first runtime.
This change fixes this issue by making sure we set the start_time to 0
for the tasks that have been scheduled in the first solver run.

Moreover, the change fixes Quincy's cost to an unscheduled aggregator node
by updating it only if the task is currently unscheduled.

Affected modules: scheduling/flow

Change-Id: I2bac1b02c707bfbcc71fec54924843fd8a85ab9c