Added cost_offset to QuincyCostModel.
Currently, costs on running arcs in QuincyCostModel quickly become
negative. This is great for from scratch Flowlessly Relax because
it pushes flow through all negative cost running arcs in a single
pass of the graph. However, this is not ideal for the cases in
which we want to demonstrate the more generalizable improvements
we've made by using incremental Relax. This change, adds a flag
to control by how much the costs should be offset.

Affected modules: scheduling/flow

Change-Id: I579861ae1e4875edc701a74ec2b37d2fe6d46c1c