Fail if Quincy cost model used with a DLM.
This change adds a CHECK to the QuincyCost model that makes its
construction fail if no DataLayerManager is present in the
knowledge base. This can happen if the KB was constructed without a DLM
argument. This CHECK also guards against the Quincy cost model being
used on a cluster that does not have the HDFS integration enabled.

Note that the accessor for the KB's DLM no longer checks if the DLM is
NULL, since we would otherwise fail early. In other words, it is now
legal for the KB to return a NULL pointer for the DLM. (However, the
const ref accessor still fails if the DLM is NULL.)

Issue: #46
Affected modules: scheduling scheduling/flow

Change-Id: I81682e102eff1645114d4878f3f34ed6bfdd1049