Workaround for CMake issue with in-source builds.
When building external dependencies inside the source tree, CMake gets
confused on incremental builds (at which point the dependency's source
tree is dirty). To work around this issue, we disable dynamic update
steps for the in-source build dependencies (cs2 and pb2json). In the
long run, we should port them to support out-of-tree builds.

Change-Id: Ib192de6899ef705f9f13ec4ad6ac863f4bfc9ba3