Disambiguate example target names on in-tree build
When building within the source tree, the names of the executables in
the examples/ directory clashed with the names of the subdirectories
containing the source for the corresponding examples. This change moves
the binaries into the per-example directories to avoid this clash.

This led to our Docker build failing (see #45), since it is an in-tree
build.

Issue: #45
Affected modules: examples

Change-Id: I86588d4ac43dfddce147e2c32ac5d4b17057587e