Update patch set 1
Patch Set 1:

I think a more elegant way to do this is to use a flag validator. You can check if hdfs_.. is enabled when we're using the Quincy cost model. In this way Firmament will stop right at the beginning at the flag validation stage. However, let's discuss this.

Patch-set: 1
Label: Verified=0