Update patch set 1
Patch Set 1:

The Mesos support was already in cluster_mix_network. However, it wasn't working correctly for jobs with more than one task. The change in cluster_mix_network just fixes that.

I think moving them to firmament experiments is a good idea. However, we would have to see how easy it is to do that given the dependency on the various protobufs.

Patch-set: 1
Label: Code-Review=0