Port code to protobuf3 APIs.
This change ports all C++ code to the new protobuf3 interfaces. They are
mostly the same as in protobuf2, but there are two differences that had
breaking effects on our code:

 1) Singular numeric fields no longer expose has_<field>() method, but
    are always initialized to the default value (zero). This requires us
    to compare against the default zero value instead of checking
    whether an optional field exists (all fields are optional in
    protobuf3).

 2) Singular string fields likewise no longer expose a has_<field>()
    method either, and are instead initialized to the empty string. This
    requires us to change all checks for these fields to compare against
    the empty string instead. Note that this requires the condition to
    be negated: previously, we checked for has_<field>() being TRUE, and
    now we need to check for <field>().empty() being FALSE (and vice
    versa).

The good news is that these changes tend to simplify our code: in
several places, conditionals can be omitted because default
initialization values have the right effect (e.g., with numeric fields
being set to zero).

Change-Id: I291e748f869d5a05c0f98c205b8b1fd9bf958261