Update patch set 2
Patch Set 2:

(1 comment)

I moved the flag, but did not yet add support for it in GoogleTraceLoader. I believe this is best done in a separate change, since it will require a change to the LoadTaskUtilizationStats() interface (which must take a reference to the task_runtime map, so that it can look up the real runtime). However, this commit is still consistent: if using the GoogleTraceLoader with --task_duration_oracle, task durations will simply be returned as zero.

Patch-set: 2