Task duration oracle support in GoogleTraceLoader.
This commit adds support for the GoogleTraceLoader to pass accurate
task runtime information into the KB. This is useful for evaluating
scheduler performance in an idealized "oracle" scenario where all
runtimes are known ahead of time. We already supported it in the
synthetic trace loader, but not in the GoogleTraceLoader.

The change necessitated a slight change to the interface for
LoadTaskRuntimeStats: it now also receives a reference to the (already
populated) task runtime map.

Affected modules: sim

Change-Id: I9c9f07d8d247a7d16261a5bf95fb96efe85174fa