Fix memory leak in SimpleScheduler.
This fixes a fairly egregious memory leak in the `SimpleScheduler`
implementation: for every resource considered, we would leak
sizeof(ResourceID_t) bytes of memory on every scheduling decision!
The effect of this becomes obvious when running a large trace (e.g., the
Google trace) in the simulator using the `SimpleScheduler`.

Affected modules: scheduling/simple

Change-Id: If5b6afd19753442bc380a57d29a3015482bf5944