standardizes job descriptory query on submit
it turns out the job_submit.py script was broken on master (including
the example command in the readme):
```
$ python job_submit.py localhost 8080 /bin/sleep 60
```

this is because as of commit:
0b8ceea527c820665ef99c98e709c30b537e21d2
Where job_submit.py started posting using the `jd` parameter, meanwhile
the handler + other scripts continued using/checking for (test).

this commit changes all instances of the `test` parameter to `jd`,
since it seems like a better parameter name in the long run, and
so the original job_submit script works.