Update patch set 1
Patch Set 1: Code-Review+1

LGTM in principle.

Will this always add padding (i.e., blank space) to the top or just when space is tight as in the screenshots? I think the former, which is not ideal: if I have a wide screen, I might not want to waste 50px at the top if the title bar doesn't wrap.

Patch-set: 1
Label: Code-Review=+1