fixes/adds json view templates.
this commit adds two json views for http:
   - json_job_status
   - json_task_status

this allows someone to hit the api and find info about the task/job
in json so one doesn't have to resort to html scraping.

next this commit fixes the json output in `json_job_completion`
which was actually responding with text, and not an actual json
object.