Update patch set 1
Patch Set 1: Verified+1 Code-Review+1

(1 comment)

This seems fine in principle, but let's briefly discuss the use case: we previously only had
 GenerateJobID()
and
 GenerateJobID(uint64_t)
because the former ensures we always generate globally unique JobIDs, while the latter is used for deterministic job IDs in the trace-driven simulator. I'm not against adding a way to compute deterministic JobIDs, but it should meet a requirement that cannot be met with the (arbitrarily configurable) job name.

(In other words, there should be a good reason to relax the assumption that JobID_t values are globally unique, because lots of places in the code certainly make that assumption.)

Patch-set: 1
Reviewer: Malte Schwarzkopf <1004901@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=+1