Only clean task state in TaskRemoval.
In the prior scheduler service API we didn't have support
for task removal. As a result, we would clean the task
state and job state when a task/all of a job's task completed.
However, users may want to have info about tasks in the system
even after they complete. This change updates the service such
that task state is only deleted when TaskRemoval is explicitly
called.

Affected modules: scheduling scheduling/flow sim

Change-Id: I127639af670523d5d37cfb036394c59bbaf697bd