Update patch set 1
Patch Set 1:

I followed the same semantics as the ones in Kubernetes. My feeling is that we would need an additional flag for tasks/jobs to indicate what we should do in such situations. For some tasks it may be worth leaving the second task instance to run to completion despite the first instance completing while the network partition happened, but for other types of tasks this is undesirable.
As it stands, we just always leave both running and we handle their completion events. We do not fail on any assertion.

Patch-set: 1