Fixed bug in handling simulated synthetic task.
In the Google trace we introduce a synthetic task at time 1s
in order to trigger a second scheduler run which pushes flow
through the running arcs created after the first scheduler run.
However, we do not advance the time for both scheduler runs.
As a result, this synthetic task can end up with a start time
smaller than submit time. This change fixes this bug.

Affected modules: sim

Change-Id: I2e824e50e3b94041a037a5de9051e92244354f07